apps/agents/markitdown-agent/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
python-multipart>=0.0.9,<1
pydantic>=2.8,<3
pydantic-settings>=2.4,<3
# Use `markitdown` for DOCX/PDF/PPTX/XLSX/HTML/EPUB (the common cases).
# Switch to `markitdown[all]` if you need image OCR or audio transcription
# (pulls in onnxruntime + ML deps; ~1GB install).
markitdown>=0.1.0,<1
pypandoc-binary>=1.13,<2
